day07: Untested-but-typechecked sketch of the amplifier loop.
Yes, I could use the single-stepping I already have to run whichever Computer
got input last.  I could even add the ability for the Device to cause a
simulation exit and cooperatively schedule that way.  And Rust has fancy
async I/O stuff now that could be used (except that would probably mess
up the types for sync use).

But spinning up a ton of threads should also work.  And back in the old
days, when Rust still had most of its Erlang influence, that was going
to be the preferred solution, but here we are.