Iterator::count exists.
I was *sure* there was a method for that, but it's been a while since
I've used Rust and forgot the name and didn't see it when scrolling
through the docs, until now.  I'd been thinking, maybe it's just not that
useful outside of toy problems and they didn't include it?  But no.