day12: Part 2, in which there is subtlety and an unanswered question.
The key insight is that the dynamics for each coordinate depend only on
that coordinate, so the problem can be split into 3, with the hope that
each one has a manageable small cycle and the large cycle for the whole
system is because the periods are close to relatively prime.

The unanswered question is whether the cycle is guaranteed to include
the initial state.  In the general case of cycle finding there can be a
lead-in before the states start repeating (see Pollard's "rho" algorithm
and the imagery of the tail vs. the loop, and see also "Garden of Eden"
states in cellular automata), and I allowed for it, but it doesn't seem
to be the case for the three initial conditions given here.  However, I
haven't tested this with any other states.