day13: Interactive mode, to check out some of the edge cases.
The input, like the comment says, is line-based because I'm too lazy
to put the tty into cbreak mode.  (I'm sure there's a crate for it.)
Yes, program input is still stdin; opening /dev/tty isn't hard, and the
Rust stdlib's stdin/stdout can't be used as a boxed abstract I/O thing
anyway because of the lock borrowing stuff.