day13: A simple bot.
The "algorithm" is just to move the paddle towards the ball if they're
not in the same column.  It is possible to watch the ball velocity and
control where it hits the paddle to manipulate its dx, but the play
area is 17x42 (relatively prime dimensions) so the ball should (and in
practice does) hit everything if it just bounces wherever.