day22: And, lo, part 2.
So I was thinking there had to be a way to simplify the stack of
permutations; like, rev+rev=id, rev+cut(n)=cut(-n)+rev, etc., and then
I realized they were all linear functions of the input (modulo the deck
size).  That space is closed under composition, so the entire input boils
down to one `y = mx + b`, and stacking a huge number of copies can be
done with the usual recursive halving/doubling thing.

(I'm using additive notation (add, multiply, negate) rather than
multiplicative (multiply, exponentiate, invert) because people on
Cryptography Twitter have opinions about it.  Granted, that's more about
typeset papers not being ridiculous with superscripts, but whatever.
And now that I actually looked it up the math convention is to use
additive notation iff the group is Abelian... which this is not.  Oops.)

An extra subtlety (which I missed at first, and submitted a wrong
answer) is that part 1 wants the position of a card but part 2 wants
the card of a position.  Which means inverting (negating, whatever) the
transformation with help from Extended Euclid.